+++ lib/amount_input_formatter.dart
@@ -1,5 +1,6 @@
+/// This package
 library amount_input_formatter;
 
-export 'src/double_extension.dart';
-export 'src/amount_input_formatter.dart';
-export 'src/number_formatter.dart';
+export 'src/amount_input_formatter.dart' show AmountInputFormatter;
+export 'src/number_formatter.dart' show NumberFormatter;
+export 'src/text_controller_extension.dart' show TextControllerExtension;