+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 0.1.0
 
-* TODO: Describe initial release.
+* # Initial package release.
+* Added `AmountInputFormatter` to use with `TextField` widget.
+* Added `NumberFormatter` to use for raw number parsing and formatting.
+* Added `TextEditingController` extension with methods `setAndFormatText` and `syncWithFormatter`.